requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
fernet==1.0.1
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
orjson==3.5.1
pip-autoremove==0.9.1
promise==2.3
pyaes==1.6.1
pydantic==1.8.1
python-dotenv==0.17.0
python-engineio==4.0.1
python-multipart==0.0.5
python-socketio==5.1.0
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
wsproto==1.0.0
